Create 5.3 Iterative scaling script
Line
Description
1...4
This should be familiar by now
5
Prompt the user to pick a single curve object, we're allowing preselection.
6
Check that the user picked an object, and that its id was written to our curve_id variable. Exit if not.
8
Retrieve the current curve length. This function should not fail, no need to check for Null.
10
Prompt the user for a length limit value. The value must be chosen between the current curve length and 1% of the current curve length. We're setting the default to half the current curve length.
13
Start a While... loop
14
This is the break-away conditional. If the curve length no longer exceeds the preset limit, the break statement will take us directly to line 20.
15
If the length of the curve did exceed the preset limit, this line will be executed. The rs.ScaleObject()
method takes four arguments, the last one of which is optional. We do not override it. We do need to specify which object we want rescaled (curve_id), what the center of the scaling operation will be ((0,0,0); the world origin) and the scaling factors along x, y and z (95% in all directions).
19
This line ends all indented code, which instructs the interpreter to go back to line 13
20
Eventually all curves will become shorter than the limit length and the While... loop will abort. We print out a message to the command line informing the user of the new curve length.