Create 6.4 An Arrow
Line
Description
1
Standard function declaration. The function takes two arguments, if the first one does not represent a proper vector array the function will not do anything, if the second one does not represent a proper point array the function will draw the vector from the world origin.
2
Declare and compute the coordinate of the arrow tip. This will potentially fail if ptBase or vecDir are not proper arrays. However, the script will continue instead of crash due to the exception handling.
3
Here we are calling the RhinoScript method rs.AddLine() and we're storing the return value
directly into the line variable. There are three possible scenarios at this point:
1.
The method completed successfully
2.
The method failed, but it didn't crash
3.
The method crashed
In the case of scenario 1, the line variable now contains the object ID for a newly added line object. This is exactly what we want the function to return on success.
In case of scenario #2, the line variable will be set to None. The last option means that there was no return value for AddLine() and hence line will also be None.
4
Check for scenario 2 and 3, and if they did not occur, go ahead and and add an arrow head using the CurveArrows method. If they did, this method will not be exectuted, and the script simply does not returns None.