Create 6.5 An AddVector() example
Line
Description
8
The listpoints variable is a list which contains all the coordinates of a pointcloud object. This is an example of a nested list (see paragraph 6.6).
12
The variable point, which is taken from the listpoints variable, contains an array of three doubles; a standard Rhino point definition. We use that point to construct a new vector definition which points from the Base point to arrPoints(i).
14
The rs.VectorCrossProduct() method will return a vector which is perpendicular to vecBase and the world z-axis. If you feel like doing some homework, you can try to replace the hard-coded direction ([0,0,1]) with a second variable point a la base_point.
15
rs.VectorCrossProduct() will fail if one of the input vectors is zero-length or if both input vectors are parallel. In those cases we will not add a vector to the document.
16 & 17
Here we make sure the vecdir vector is two units long. First we unitize the vector, making it one unit long, then we double the length.
25
Finally, place a call to the AddVector() function we defined on page 40. If you intend to run this script, you must also include the AddVector() function in the same script.