Create 8.3.1 Colourful Sprials
The following script will add 500 colored points to the document, all of which are sampled at regular intervals across the R1 parameter space of a curve object:
For no good reason whatsoever, we'll start with the bottom most function:
Line
Description
24
Standard out-of-the-box function declaration which takes a single double value. This function is
supposed to return a colour which changes gradually from blue to red as parameter changes from zero to one. Values outside of the range {0.0~1.0} will be clipped.
25
The red component of the colour we're going to return is declared here and assigned the naive value of 255 times the parameter. Colour components must have a value between and including 0 and 255. If we attempt to construct a colour with lower or higher values a run-time error will spoil the party.
26…27
Here's where we make sure the party can continue unimpeded.
28
Compute the colour gradient value. If parameter equals zero we want blue (0,0,255) and if it equals one we want red (255,0,0). So the green component is always zero while blue and red see-saw between 0 and 255.
Line
Description
14
Function declaration.
15…16
Get the curve domain and check for Null. It will be Null if the ID does not represent a proper curve object. The Rhino.CurveDomain() method will return an array of two doubles which indicate the minimum and maximum t-parameters which lie on the curve.
18
Translate the unitized R1 coordinate into actual domain coordinates.
19
Evaluate the curve at the specified parameter. Rhino.EvaluateCurve() takes an R1 coordinate and returns an R3 coordinate.
21
Add the point, it will have default attributes.
22
Set the custom colour. This will automatically change the color-source attribute to By Object.